# globally cache small Term variables