Store `Var i []` for `i < 128` statically. In a72399c04ef9701e8713062fef20e240e4c5cd03, small variables are only cached on deserialization, now they are cached pretty much everywhere, by turning the original `Var` constructor into a pattern synonyms which looks up from the cache. This mildly decreases total memory usage for the `std-lib-test`, decreases total allocations from 612 GB to 606 GB and leaves runtime roughly the same. 

Here's a small sample of `std-lib-test` total memory usages:

- On commit b95a2ca4de00f5a7c35b7edc4e0183970b68f999, before serialization changes:
  - Runs (MB): 2960, 3000, 2966, 2949, 3151, 2946
  - Max: 3151
  - Average: 2995.34
- On commit a72399c04ef9701e8713062fef20e240e4c5cd03:
   - Runs (MB): 3235,3132,3118,3142,2794,3119
   - Max: 3235
   - Average: 3090
On this PR:
   - Runs (MB): 3083,2771,2985,2677,3105,2742
   - Max: 3105
   - Average: 2894