@lawcho 2025-08-25
This PR is missing some updates to test cases:

```shellsession
/tmp/tmp.VRwcnrAASG/agda$ git switch --detach fe2e0c801869f9f64893e3f196bd6e5b22c35fdd
HEAD is now at fe2e0c8 globally cache small Term variables (#8073)
/tmp/tmp.VRwcnrAASG/agda$ nix develop .#fail
/tmp/tmp.VRwcnrAASG/agda$ agda-tests -p DebugWith
Using JS node binary at /nix/store/a1kxazxkgw7mjbjgisvah95p1r3n5ykl-nodejs-22.16.0/bin/node
all
  Fail
    DebugWith: Golden value differs from actual value.
========================================

diff --git a/tmp/nix-shell.luN5ph/DebugWith195393-0.golden b/tmp/nix-shell.luN5ph/DebugWith195393-1.actual
index beda7a2..7b2b56e 100644
--- a/tmp/nix-shell.luN5ph/DebugWith195393-0.golden
+++ b/tmp/nix-shell.luN5ph/DebugWith195393-1.actual
@@ -32,7 +32,7 @@ checkWithFunction
   perm'  = x0,x1,x2,x3,x4 -> x1,x3,x0,x2,x4
   perm   = x0,x1,x2,x3,x4 -> x0,x1,x2,x3,x4
   fperm  = x0,x1,x2,x3,x4 -> x1,x3,x0,x2,x4
-  withSub= Var 0 [] :# (Var 4 [] :# (Var 1 [] :# (Var 5 [] :# (Var 2 [] :# Wk 6 IdS))))
+  withSub= Var' 0 [] :# (Var' 4 [] :# (Var' 1 [] :# (Var' 5 [] :# (Var' 2 [] :# Wk 6 IdS))))
 with function call DebugWith.with-34 b f (f b) trash1 trash2 trash3
 creating with display form...
 created with display form

========================================
               FAIL (0.00s)
               Golden value does not match actual output.

1 out of 1 tests failed (0.04s)
/tmp/tmp.VRwcnrAASG/agda$ 
exit
/tmp/tmp.VRwcnrAASG/agda$ git switch --detach a72399c04ef9701e8713062fef20e240e4c5cd03
Previous HEAD position was fe2e0c8 globally cache small Term variables (#8073)
HEAD is now at a72399c WIP: medium-sized serialization optimizations (#8065)
/tmp/tmp.VRwcnrAASG/agda$ nix develop .#fail
/tmp/tmp.VRwcnrAASG/agda$ agda-tests -p DebugWith
Using JS node binary at /nix/store/a1kxazxkgw7mjbjgisvah95p1r3n5ykl-nodejs-22.16.0/bin/node
all
  Fail
    DebugWith: OK (0.08s)

All 1 tests passed (0.08s)
/tmp/tmp.VRwcnrAASG/agda$ 
```