@AndrasKovacs 2025-08-25
Thanks, I'll look at it.
BTW, how can I make all tests run when I make a PR? Sometimes I only get some "checks", how does this work?