@lawcho 2025-08-25
The GitHub actions CI is currently broken https://github.com/agda/agda/pull/8071#issuecomment-3194152611

I run the tests manually with `nix build .#test` (which stops at the first failing test)

See #8036 for tips on running the nixified test suite (e.g. finer-grained tests).

If you find a test that appears to fail because of the nix wrapper (rather than Agda having bugs), please ping me and I'll try to fix it.